+++ src/main/java/com/cliptripbe/feature/bookmark/infrastructure/BookmarkRepository.java
@@ -1,7 +1,7 @@
 package com.cliptripbe.feature.bookmark.infrastructure;
 
 import com.cliptripbe.feature.bookmark.domain.entity.Bookmark;
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.EntityGraph;
