+++ src/main/java/com/cliptripbe/feature/place/application/PlaceService.java
@@ -19,9 +19,8 @@ import com.cliptripbe.feature.place.dto.request.PlaceSearchByKeywordRequest;
 import com.cliptripbe.feature.place.dto.response.PlaceAccessibilityInfoResponse;
 import com.cliptripbe.feature.place.dto.response.PlaceListResponse;
 import com.cliptripbe.feature.place.dto.response.PlaceResponse;
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import com.cliptripbe.feature.user.domain.type.Language;
-import com.cliptripbe.global.response.exception.CustomException;
 import com.cliptripbe.infrastructure.google.service.GooglePlacesService;
 import com.cliptripbe.infrastructure.kakao.service.KakaoMapService;
 import com.cliptripbe.infrastructure.s3.S3Service;
