+++ src/main/java/com/cliptripbe/feature/schedule/application/ScheduleRegister.java
@@ -2,7 +2,7 @@ package com.cliptripbe.feature.schedule.application;
 
 import com.cliptripbe.feature.schedule.domain.entity.Schedule;
 import com.cliptripbe.feature.schedule.infrastructure.ScheduleRepository;
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
