+++ src/main/java/com/cliptripbe/feature/schedule/domain/entity/Schedule.java
@@ -1,6 +1,6 @@
 package com.cliptripbe.feature.schedule.domain.entity;
 
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
