+++ src/main/java/com/cliptripbe/feature/user/api/UserController.java
@@ -2,10 +2,10 @@ package com.cliptripbe.feature.user.api;
 
 import static com.cliptripbe.global.constant.Constant.API_VERSION;
 
-import com.cliptripbe.feature.user.api.dto.request.UserSignInRequestDto;
-import com.cliptripbe.feature.user.api.dto.request.UserSignUpRequestDto;
-import com.cliptripbe.feature.user.api.dto.response.UserInfoResponse;
-import com.cliptripbe.feature.user.api.dto.response.UserLoginResponse;
+import com.cliptripbe.feature.user.dto.request.UserSignInRequest;
+import com.cliptripbe.feature.user.dto.request.UserSignUpRequest;
+import com.cliptripbe.feature.user.dto.response.UserInfoResponse;
+import com.cliptripbe.feature.user.dto.response.UserLoginResponse;
 import com.cliptripbe.feature.user.application.UserService;
 import com.cliptripbe.global.response.ApiResponse;
 import com.cliptripbe.global.response.type.SuccessType;
@@ -27,7 +27,7 @@ public class UserController implements UserControllerDocs {
     @Override
     @PostMapping("/sign-up")
     public ApiResponse<?> signUp(
-        @RequestBody UserSignUpRequestDto signUpDto) {
+        @RequestBody UserSignUpRequest signUpDto) {
         UserInfoResponse studentSignUpResponse = userService.signUp(signUpDto);
         return ApiResponse.success(SuccessType.CREATED, studentSignUpResponse);
     }
@@ -35,9 +35,9 @@ public class UserController implements UserControllerDocs {
     @Override
     @PostMapping("/sign-in")
     public ApiResponse<UserLoginResponse> signIn(
-        @RequestBody UserSignInRequestDto userSignInRequestDto
+        @RequestBody UserSignInRequest userSignInRequest
     ) {
-        UserLoginResponse userLoginResponse = userService.userSignIn(userSignInRequestDto);
+        UserLoginResponse userLoginResponse = userService.userSignIn(userSignInRequest);
         return ApiResponse.success(SuccessType.OK, userLoginResponse);
     }
