--- src/main/java/com/cliptripbe/feature/user/api/dto/request/UserSignInRequestDto.java
@@ -1,8 +1,0 @@
-package com.cliptripbe.feature.user.api.dto.request;
-
-public record UserSignInRequestDto(
-    String email,
-    String password
-) {
-
-}
