+++ src/main/java/com/cliptripbe/feature/user/dto/request/UserSignInRequest.java
@@ -1,0 +1,8 @@
+package com.cliptripbe.feature.user.dto.request;
+
+public record UserSignInRequest(
+    String email,
+    String password
+) {
+
+}
