+++ src/main/java/com/cliptripbe/feature/user/infrastructure/UserRepository.java
@@ -1,6 +1,6 @@
 package com.cliptripbe.feature.user.infrastructure;
 
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
