+++ src/main/java/com/cliptripbe/feature/video/application/VideoPlaceExtractFacade.java
@@ -1,15 +1,12 @@
 package com.cliptripbe.feature.video.application;
 
-import static com.cliptripbe.global.response.type.ErrorType.CHATGPT_NO_RESPONSE;
-import static com.cliptripbe.global.response.type.ErrorType.KAKAO_MAP_NO_RESPONSE;
-
 import com.cliptripbe.feature.place.dto.PlaceDto;
 import com.cliptripbe.feature.place.application.PlaceService;
 import com.cliptripbe.feature.place.application.PlaceTranslationService;
 import com.cliptripbe.feature.place.domain.entity.Place;
 import com.cliptripbe.feature.schedule.application.ScheduleService;
 import com.cliptripbe.feature.schedule.domain.entity.Schedule;
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.entity.User;
 import com.cliptripbe.feature.user.domain.type.Language;
 import com.cliptripbe.feature.video.domain.entity.Video;
 import com.cliptripbe.feature.video.dto.request.ExtractPlaceRequest;
