+++ src/main/java/com/cliptripbe/global/auth/security/CustomerDetailsService.java
@@ -1,7 +1,7 @@
 package com.cliptripbe.global.auth.security;
 
-import com.cliptripbe.feature.user.application.UserLoader;
-import com.cliptripbe.feature.user.domain.User;
+import com.cliptripbe.feature.user.domain.service.UserLoader;
+import com.cliptripbe.feature.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
